Move mesh.(h|cpp) into a meshes subfolder
