+++ simulant/loaders/obj_loader.cpp
@@ -20,7 +20,7 @@
 
 #include "obj_loader.h"
 
-#include "../mesh.h"
+#include "../meshes/mesh.h"
 #include "../resource_manager.h"
 #include "../shortcuts.h"
 #include "../resource_locator.h"
