+++ simulant/procedural/mesh/cube.h
@@ -21,7 +21,7 @@
 
 #include "../constants.h"
 
-#include "../../mesh.h"
+#include "../../meshes/mesh.h"
 
 namespace smlt {
 namespace procedural {
